pyproject.toml
[project]
name = "goldybot4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.3",
    "fastapi>=0.115.9",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "langgraph-prebuilt>=0.2.2",
    "langgraph-supervisor>=0.0.27",
    "praw>=7.8.1",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
